Handle previously invalid replicas
t:Node.replicas/0 is defined as pos_integer().

Previously if a value less than 1 was provided the library would use
this in a range and the results would often be wrong an counter
intuitive.  For instance, replicas = 0 would result in two replicas,
[1, 0].  Replicas = -1 would result in three replicas, [1, 0, -1]

Guard has been added so that replicas < 1 expand into the empty set, [].